backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
openai==1.58.1
resend==2.7.0
fpdf2==2.8.2
apscheduler==3.11.0
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dateutil==2.9.0
jinja2==3.1.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
email-validator==2.3.0
